+++ src/registrar/forms/portfolio.py
@@ -3,10 +3,9 @@
 import logging
 from django import forms
 from django.core.validators import RegexValidator
-from django.core.validators import MaxLengthValidator
 
+from registrar.forms.utility.fields import MaxLengthFirstEmailField
 from registrar.forms.utility.combobox import ComboboxWidget
-from registrar.validations import EMAIL_MAX
 from registrar.models import (
     PortfolioInvitation,
     UserPortfolioPermission,
@@ -442,21 +441,6 @@ class PortfolioInvitedMemberForm(BasePortfolioMemberForm):
     class Meta:
         model = PortfolioInvitation
         fields = ["roles", "additional_permissions"]
-
-
-class MaxLengthFirstEmailField(forms.EmailField):
-    """Email field that returns max-length errors before format errors."""
-
-    def __init__(self, *args, email_max_length=EMAIL_MAX, email_max_length_message=None, **kwargs):
-        self.email_max_length_validator = MaxLengthValidator(
-            email_max_length,
-            message=email_max_length_message or f"Email must be no more than {email_max_length} characters.",
-        )
-        super().__init__(*args, **kwargs)
-
-    def run_validators(self, value):
-        self.email_max_length_validator(value)
-        super().run_validators(value)
 
 
 class PortfolioNewMemberForm(BasePortfolioMemberForm):
